+++ src/net/evmodder/EvLib/noteblockapi/PositionSongPlayer.java
@@ -22,7 +22,7 @@ public class PositionSongPlayer extends SongPlayer{
 		for(Layer l : song.getLayerHashMap().values()){
 			Note note = l.getNote(tick);
 			if(note == null) continue;
-			p.playSound(targetLocation, Instrument.getInstrument(note.getInstrument()), (l.getVolume() * (int)volume * (int)playerVolume) / 1000000f,
+			p.playSound(targetLocation, Instrument.getInstrument(note.getInstrument()), (l.getVolume() * volume * playerVolume) / 1000000f,
 					NotePitch.getPitch(note.getKey() - 33));
 		}
 	}
